requirements.txt
opencv-python
torchreid
torchvision
torch
ultralytics
scikit-learn
Pillow